requirements.txt
# Core dependencies for all approaches
python-docx>=0.8.11
PyPDF2>=3.0.0
google-genai>=0.3.0

# For Approaches 1, 2, 3 (Embedding-based)
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# For Approach 2 (NLI) - Additional dependencies
transformers>=4.30.0
torch>=2.0.0
tqdm>=4.65.0

# Optional: For development and testing
# jupyter>=1.0.0
# pytest>=7.0.0
# matplotlib>=3.7.0
